Update CalendarSync.gs
Now includes the ability to define colors for holds and travel buffers. Also, rewritten to use event tags so that it's more stable. Sometimes I end up with 2 events at the same time with the same name (I made a hold, then someone sent the formal invite), so this way it will be able to distinguish those events.